Move testcases from brouter-server to their modules
