+++ brouter-core/src/test/resources/testtrack0.gpx
