+++ brouter-mapaccess/src/test/java/btools/mapaccess/IntegrityCheckTest.java
@@ -1,4 +1,4 @@
-package btools.server;
+package btools.mapaccess;
 
 import java.io.File;
 import java.net.URL;
